Addition of SMS, update to serialPort and start of camera
The serialPort now allows for either comName to be used or an object where the key=val pairing will be used to dynamically connect. This is useful for arduinos where PC, macOS and Linux handle the names differently.

An SMS module has been added that takes the already easy to use Twilio module. There is also an example to demonstrate the usage.

Instead of having optional installations, all the potential NPM installs have been added to ensure that there is no confusion as to which need to be installed for individual Æther modules. They won’t be loaded into the application unless they are needed but having them installed is a minimal addition for ensuring that none are missing for on-the-fly additions.

There is also an upcoming camera module that is going to be added that. There is now a reserved variable for this.

Signed-off-by: Matthew Potter <mdpotter@gmail.com>